egra_egma_datavizir/requirements.txt
streamlit
pandas
numpy
plotly
scipy
python-docx
openpyxl
statsmodels
kaleido
transformers
torch
python-dotenv
anthropic
matplotlib
seaborn